ref: extract context folder in services
