+++ randomizer/src/components/Input/Input.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useUserListContext from '../../contexts/UsersList';
+import useUserListContext from '../../services/contexts/UsersList';
 import { DataStr } from './types';
 
 import './index.scss';
