+++ randomizer/src/services/contexts/UsersList/index.tsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import Context from './Context';
 import Provider from './Provider';
-import { IUserListContext } from '../../interfaces';
+import { IUserListContext } from '../../../interfaces';
 
 const useUserListContext = (): IUserListContext => useContext(Context);
